requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
albumentations==1.3.0
appdirs==1.4.4
async-timeout==3.0.1
attrs==22.2.0
autopep8==1.6.0
azure-common==1.1.26
azure-core==1.10.0
azure-storage-blob==12.3.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beartype==0.14.0
biopython==1.76
brotlipy==0.7.0
cairocffi==1.4.0
CairoSVG==2.6.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.2.1
colored==1.4.4
cryptography==39.0.1
cssselect2==0.7.0
cycler==0.11.0
datasets==2.10.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.4
filelock==3.9.0
fonttools==4.38.0
fsspec==2023.1.0
fuzzysearch==0.7.3
fuzzywuzzy==0.18.0
genalog==0.1.0
gitdb==4.0.10
GitPython==3.1.31
html5lib==1.1
huggingface-hub==0.14.1
humanfriendly==10.0
humanize==4.6.0
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
ipykernel==6.15.0
ipython==7.34.0
ipywidgets==8.0.6
isodate==0.6.1
jedi==0.18.2
Jinja2==2.11.1
joblib==1.2.0
jupyter-client==7.0.6
jupyter_core==4.11.1
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
MarkupSafe==2.0.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msrest==0.6.21
multidict==4.7.6
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==2.6.3
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.2.0.34
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.0.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
pipreqs==0.4.13
prompt-toolkit==3.0.37
protobuf==4.22.0
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==11.0.0
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyphen==0.13.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.11.0
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==19.0.2
qudida==0.0.4
regex==2023.5.5
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
sacremoses==0.0.53
scikit-image==0.16.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.9.0
setproctitle==1.3.2
setuptools==65.6.3
six==1.16.0
slurm-gpustat==0.0.12
smmap==5.0.0
submitit==1.4.5
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
torch==1.13.1
torchvision==0.14.1
tornado==6.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.18.0
typing_extensions==4.5.0
urllib3==1.26.14
wandb==0.13.10
wcwidth==0.2.6
WeasyPrint==51
webencodings==0.5.1
wheel==0.38.4
widgetsnbextension==4.0.7
xxhash==3.2.0
yarg==0.1.9
yarl==1.8.2
zipp==3.15.0
